# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateCategoryRequest:
    properties:
      color:
        example: '#3498db'
        type: string
      description:
        example: Gastos com viagens e turismo
        type: string
      icon:
        example: ✈️
        type: string
      name:
        example: Viagens
        type: string
    required:
    - name
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
...
  handler.UpdateCategoryRequest:
    properties:
      color:
        example: '#e74c3c'
        type: string
      description:
        example: Restaurantes e delivery
        type: string
      icon:
        example: "\U0001F355"
        type: string
      name:
        example: Alimentação Fora
        type: string
    type: object
...
host: 147.185.221.212:61489
info:
  contact:
    email: suporte@exemplo.com
    name: Equipe de Desenvolvimento
    url: https://github.com/Pmmvito/Golang-Api-Exemple
  description: "API REST para autenticação JWT, análise inteligente de notas fiscais
    de supermercado e gestão de categorias\n\n## \U0001F680 Funcionalidades Principais:\n-
    \U0001F510 **Autenticação JWT** com tokens de 7 dias\n- \U0001F916 **Análise de
    Notas com IA** usando Google Gemini (2-3s)\n- \U0001F4CA **Categorização Automática**
    de produtos de supermercado\n- \U0001F4DD **CRUD Completo** de recibos e categorias\n-
    \U0001F50D **Filtros Avançados** por categoria, data e valor\n- ⚡ **Respostas
    Otimizadas** (55% menores)\n\n## \U0001F4D6 Como Usar:\n1. Registre-se em `/register`\n2.
    Faça login em `/login` para obter o token JWT\n3. Clique em **Authorize** \U0001F513
    (cadeado verde) e cole: `Bearer SEU_TOKEN`\n4. Agora você pode testar todos os
    endpoints protegidos!\n\n## \U0001F6D2 Categorias de Supermercado (21 categorias
    distintas):\n**Básicos:** Grãos e Cereais, Massas, Padaria\n**Proteínas:** Carnes
    e Proteínas, Frios e Embutidos\n**Laticínios:** Leite, Queijos, Iogurtes, Manteiga\n**Frescos:**
    Frutas e Vegetais\n**Bebidas:** Bebidas (não alcoólicas), Bebidas Alcoólicas,
    Café e Chá\n**Congelados:** Produtos Congelados\n**Doces:** Doces e Sobremesas,
...
    Higiene Pessoal, Limpeza Doméstica, Papel e Descartáveis\n**Especiais:** Bebê
...
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: 'Get all expense categories sorted by name. Returns: 1-Alimentação,
        2-Transporte, 3-Saúde, 4-Lazer, 5-Educação, 6-Moradia, 7-Vestuário, 8-Outros'
      produces:
      - application/json
      responses:
        "200":
          description: List of categories with count
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all categories
      tags:
      - "\U0001F4C1 Categories"
  /categories/graph:
    get:
      consumes:
      - application/json
      description: Get aggregated data for each category, including item count and
        total value. Filters by date range, defaulting to the current month.
      parameters:
      - description: Start date for filtering (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date for filtering (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category graph data retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get category graph data
      tags:
      - "\U0001F4C1 Categories"
  /category:
    post:
      consumes:
      - application/json
      description: Create a new expense category for organizing receipt items
      parameters:
      - description: Category data (name is required, description/icon/color are optional)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new category
      tags:
      - "\U0001F4C1 Categories"
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: 'Delete a category permanently. WARNING: This will also remove
        the category association from all receipt items (sets categoryId to null).'
      parameters:
      - description: Category ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - "\U0001F4C1 Categories"
    get:
      consumes:
      - application/json
      description: Get details of a specific category by ID including all items that
        belong to this category
      parameters:
      - description: Category ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category details with items
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get category details
      tags:
      - "\U0001F4C1 Categories"
    patch:
      consumes:
      - application/json
      description: Update category information (name, description, icon, color). All
        fields are optional - only send what you want to update.
      parameters:
      - description: Category ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Category data to update (all fields optional)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or no fields to update
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - "\U0001F4C1 Categories"
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer" seguido do seu token JWT (obtido no login). Exemplo:
      Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"